Revert buggy upstream Multilanguage commit
Revert MultiLanguage to b0e47e7e.

Multilanguage/8aa81b5 introduced a bug where translations would
fail when the browser sends a 4 character locale code but the
plugin would set the locale code to a 2 character version. This
resulted in failure to find the correct translations and
nothing but dates would be properly translated.