Remove oc translation files
This is a bit of a weird one. Zend_Locale throws an exception
when Zend_Locale::getTranslation('oc', 'language', 'oc') is
called when getting translations of language names in the
Multilanguage plugin config page. Since this site will never be
used in Occitan, I think it's safe to just remove these files and
worry about it no more.

https://en.wikipedia.org/wiki/Occitan_language