Exhibit gallery bookmark and hover overlays
Adds markup and CSS for bookmark and hover info in exhibit image
galleries. For now it uses the FontAwesome bookmark glyph we
which is already available. We can potentially switch it to an
image (or even better, a really simple inline SVG file) if this
look isn't quite right. This way if we need to change colors or
border it is super simple.

This puts all available book location info into the on-hover
overlay with a label stating what it is.

Fixes #32.