Proper handling of embedded video in galleries
The image file associated with the item containing the Embed field
is shown in the gallery with a video file icon in the upper left.
This file is only used for the preview and not shown on the item
show page, instead showing the embedded iframe with the video.
Note: Embedded videos must have an uploaded file to display
properly in Exhibit galleries.

Fixes #31.